new works

+++ src/main/java/com/obswrldEcommerceApp/services/UserServiceImplementation.java
@@ -9,6 +9,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+
 @Service
 @RequiredArgsConstructor
 public class UserServiceImplementation implements UserServiceInterface{
@@ -17,6 +19,9 @@ public class UserServiceImplementation implements UserServiceInterface{
 
     @Override
     public UserResponse register(UserRegistrationRequest userRegistrationRequest) {
+        if (userRepositories.existsByEmail(userRegistrationRequest.getEmail())) {
+            throw new RuntimeException("Email already exists");
+        }
         String password = bCryptPasswordEncoder.encode(userRegistrationRequest.getPassword());
         User user = User.builder()
                 .name(userRegistrationRequest.getName())
@@ -25,30 +30,60 @@ public class UserServiceImplementation implements UserServiceInterface{
                 .roles(userRegistrationRequest.getRoles())
                 .address(userRegistrationRequest.getAddress())
                 .build();
-
         User savedUser = userRepositories.save(user);
-        return new UserResponse(savedUser.getUserId(), savedUser.getName(), savedUser.getEmail(),
-                savedUser.getRoles(), savedUser.getAddress());
+        return new UserResponse(
+                savedUser.getUserId(),
+                savedUser.getName(),
+                savedUser.getEmail(),
+                savedUser.getRoles(),
+                savedUser.getAddress());
     }
 
     @Override
     public UserLoginResponse login(UserLoginRequest userLoginRequest) {
-        return userRepositories.findByEmail(userLoginRequest.getEmail())
-                .filter(user -> bCryptPasswordEncoder.matches(userLoginRequest.getPassword(), user.getPassword()))
-                .map(user -> new UserLoginResponse("Login Successful", user.getUserId(), user.getRoles(),
-                        user.getEmail()))
-                .orElse(new UserLoginResponse("Invalid Password", null, null, null));
+        User user = userRepositories.findByEmail(userLoginRequest.getEmail())
+                .orElseThrow(() -> new RuntimeException("User not found"));
+        if (!bCryptPasswordEncoder.matches(userLoginRequest.getPassword(), user.getPassword())) {
+            throw new RuntimeException("Invalid password");
+        }
+        return new UserLoginResponse(
+                "Login Succesful",
+                user.getUserId(),
+                user.getRoles(),
+                user.getEmail()
+        );
     }
 
     @Override
     public UserResponse findByEmail(String email) {
         User user = userRepositories.findByEmail(email)
                 .orElseThrow(() -> new RuntimeException("User not found"));
-        return new UserResponse(user.getUserId(), user.getName(), user.getEmail(), user.getRoles(), user.getAddress());
+        return new UserResponse(user.getUserId(),
+                user.getName(),
+                user.getEmail(),
+                user.getRoles(),
+                user.getAddress()
+        );
     }
 
     @Override
     public void deleteUser(String userId){
+        if(!userRepositories.existsByEmail(userId)) {
+            throw new RuntimeException("User not found");
+        }
         userRepositories.deleteById(userId);
+    }
+
+    @Override
+    public List<UserResponse> getAllUsers() {
+        return userRepositories.findAll().stream()
+                .map(user -> new UserResponse(
+                        user.getUserId(),
+                        user.getName(),
+                        user.getEmail(),
+                        user.getRoles(),
+                        user.getAddress()
+                ))
+                .toList();
     }
 }
